Major changes to twinc . . .
Twinc now checks for updates to the main timeline.  If it finds any, it simultaneously writes those to the screen (in digest form) and serializes them as numbered files in $HOME/Twitter/inbox.  The numbered files contain Clojure forms with the twitter id as the key and a struct (containing various items from the twitter4j.Twitter object) as values.

None of this is written in stone.  There are a few functions that seem a bit verbose to me, and I'm not sure I'm following the proper Clojure idiom in all cases.  I like the idea of using human-readable lisp as the serialization format ("Store data in flat text files"), but there may be disadvantages to this.